+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.7.0
 - `new` Improved trends report graph. It should not have data cut offs where months have different length.
+- `new` Trends report now has detailed amounts for each day.
 
 ### v0.6.0
 - `new` Date formatting now includes year.